+++ sawatabi/algorithm/new.py
@@ -12,10 +12,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import apache_beam as beam
+
 from sawatabi.algorithm.abstract_algorithm import AbstractAlgorithm
 
 
 class New(AbstractAlgorithm):
+
+    @classmethod
+    def create_pipeline(cls, algorithm_options, input_fn=None, map_fn=None, solve_fn=None, unmap_fn=None, output_fn=None, pipeline_args=["--runner=DirectRunner"]):
+        algorithm_transform = (
+            "Fixed windows" >> beam.WindowInto(beam.window.FixedWindows(size=algorithm_options["window.size"]))
+            | "Add timestamp as tuple againt each window for diff detection" >> beam.ParDo(AbstractAlgorithm.WithTimestampTupleFn())
+            | "Elements in a fixed window into a list" >> beam.CombineGlobally(beam.combiners.ToListCombineFn()).without_defaults()
+            | "To a single global Window from fixed windows" >> beam.WindowInto(beam.window.GlobalWindows())
+        )
+
+        return cls._create_pipeline(algorithm_transform=algorithm_transform, algorithm_options=algorithm_options, input_fn=input_fn, map_fn=map_fn, solve_fn=solve_fn, unmap_fn=unmap_fn, output_fn=output_fn, pipeline_args=pipeline_args)
+
+
     ################################
     # Built-in functions
     ################################
